LayoutTests/imported/w3c/web-platform-tests/storage/estimate-indexeddb.https.any.worker-expected.txt


PASS estimate() method exists and returns a Promise
PASS estimate() resolves to dictionary with members
PASS estimate() shows usage increase after large value is stored
